moved community search component
